requirement.txt
streamlit
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
librosa
matplotlib
Pillow
